fix(mir): complete VecIter ownership handoffs
